% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corMLPE_class.R
\name{corMLPE}
\alias{corMLPE}
\title{Correlation structure for symmetric relational data}
\usage{
corMLPE(value = 0.1, form = ~1, fixed = FALSE, covariate = NULL)
}
\arguments{
\item{value}{Starting value for the correlation parameter}

\item{form}{A formula in the form ...}

\item{fixed}{Logical, fit model with starting value fixed}

\item{covariate}{Optional, supply a covariate}
}
\description{
Correlation structure for symmetric relational data
}

