% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corMLPE_class.R
\name{dgCMatrix_corMLPE-methods}
\alias{dgCMatrix_corMLPE-methods}
\alias{t,dgCMatrix_corMLPE-method}
\alias{[,dgCMatrix_corMLPE,index,missing,logical-method}
\alias{[,dgCMatrix_corMLPE,missing,index,logical-method}
\title{Methods for Internal `dgCMatrix_corMLPE` Objects}
\usage{
\S4method{t}{dgCMatrix_corMLPE}(x)

\S4method{[}{dgCMatrix_corMLPE,index,missing,logical}(x, i, j, ..., drop = TRUE)

\S4method{[}{dgCMatrix_corMLPE,missing,index,logical}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{A `dgCMatrix_corMLPE` object.}

\item{i}{Row index.}

\item{j}{Column index.}

\item{...}{Additional arguments passed to matrix subsetting.}

\item{drop}{Logical; passed to `[` methods.}
}
\description{
These methods preserve class identity required for `mgcv::gamm()`
compatibility when sparse matrix operations are applied.
}
