% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simulate_corMLPE_residuals}
\alias{simulate_corMLPE_residuals}
\title{Simulate data from a fitted corMLPE model}
\usage{
simulate_corMLPE_residuals(model, n = 1, seed = NULL)
}
\arguments{
\item{model}{fitted model of class 'gls' with corMLPE correlation structure}

\item{n}{number of simulations}

\item{seed}{random seed}
}
\description{
Simulate data from a fitted corMLPE model
}
